WIP: fix tests
